+++ src/mflux/models/ernie_image/model/mistral3_text_encoder/__init__.py
@@ -1,0 +1,3 @@
+from mflux.models.ernie_image.model.mistral3_text_encoder.text_encoder import Mistral3TextEncoder
+
+__all__ = ["Mistral3TextEncoder"]